Add : sample resource img, Fix : CA to libray, UI
