+++ BoneAgeChecker.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 14.0.23107.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BoneAgeChecker", "BoneAgeChecker\BoneAgeChecker.csproj", "{46AD2E11-C62A-43E1-8056-CBFF48772B52}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ContourAnalysis", "CA\ContourAnalysis.csproj", "{DDBD6CF7-2367-4084-A609-1B9C1885DC98}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -28,6 +30,18 @@ Global
 		{46AD2E11-C62A-43E1-8056-CBFF48772B52}.Release|x64.Build.0 = Release|x64
 		{46AD2E11-C62A-43E1-8056-CBFF48772B52}.Release|x86.ActiveCfg = Release|x86
 		{46AD2E11-C62A-43E1-8056-CBFF48772B52}.Release|x86.Build.0 = Release|x86
+		{DDBD6CF7-2367-4084-A609-1B9C1885DC98}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DDBD6CF7-2367-4084-A609-1B9C1885DC98}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DDBD6CF7-2367-4084-A609-1B9C1885DC98}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{DDBD6CF7-2367-4084-A609-1B9C1885DC98}.Debug|x64.Build.0 = Debug|Any CPU
+		{DDBD6CF7-2367-4084-A609-1B9C1885DC98}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{DDBD6CF7-2367-4084-A609-1B9C1885DC98}.Debug|x86.Build.0 = Debug|Any CPU
+		{DDBD6CF7-2367-4084-A609-1B9C1885DC98}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DDBD6CF7-2367-4084-A609-1B9C1885DC98}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DDBD6CF7-2367-4084-A609-1B9C1885DC98}.Release|x64.ActiveCfg = Release|Any CPU
+		{DDBD6CF7-2367-4084-A609-1B9C1885DC98}.Release|x64.Build.0 = Release|Any CPU
+		{DDBD6CF7-2367-4084-A609-1B9C1885DC98}.Release|x86.ActiveCfg = Release|Any CPU
+		{DDBD6CF7-2367-4084-A609-1B9C1885DC98}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
