+++ BoneAgeChecker/BoneAgeChecker.csproj
@@ -117,7 +117,6 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Contour.cs" />
     <Compile Include="InitView.cs">
       <SubType>Form</SubType>
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
@@ -135,14 +134,12 @@
     <Compile Include="ImageProcessor.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Template.cs" />
     <Compile Include="TemplateEditor.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="TemplateEditor.designer.cs">
       <DependentUpon>TemplateEditor.cs</DependentUpon>
     </Compile>
-    <Compile Include="TemplateFinder.cs" />
     <Compile Include="TemplateGenerator.cs" />
     <EmbeddedResource Include="InitView.resx">
       <DependentUpon>InitView.cs</DependentUpon>
@@ -190,6 +187,43 @@
     <Content Include="Emgu.Util.dll">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
+    <None Include="img\f0y.jpg" />
+    <None Include="img\f1y.jpg" />
+    <None Include="img\f2y.jpg" />
+    <None Include="img\f3y.jpg" />
+    <None Include="img\f4y.jpg" />
+    <None Include="img\f5y.jpg" />
+    <None Include="img\f6y.jpg" />
+    <None Include="img\f7y.jpg" />
+    <None Include="img\f8y.jpg" />
+    <None Include="img\f10y.jpg" />
+    <None Include="img\f11y.jpg" />
+    <None Include="img\f12y.jpg" />
+    <None Include="img\f13y.jpg" />
+    <None Include="img\f14y.jpg" />
+    <None Include="img\f15y.jpg" />
+    <None Include="img\f16y.jpg" />
+    <None Include="img\f17y.jpg" />
+    <None Include="img\f18y.jpg" />
+    <None Include="img\m0y.jpg" />
+    <None Include="img\m1y.jpg" />
+    <None Include="img\m2y.jpg" />
+    <None Include="img\m4y.jpg" />
+    <None Include="img\m5y.jpg" />
+    <None Include="img\m6y.jpg" />
+    <None Include="img\m7y.jpg" />
+    <None Include="img\m8y.jpg" />
+    <None Include="img\m9y.jpg" />
+    <None Include="img\m10y.jpg" />
+    <None Include="img\m11y.jpg" />
+    <None Include="img\m12y.jpg" />
+    <None Include="img\m13y.jpg" />
+    <None Include="img\m14y.jpg" />
+    <None Include="img\m15y.jpg" />
+    <None Include="img\m16y.jpg" />
+    <None Include="img\m17y.jpg" />
+    <None Include="img\m18y.jpg" />
+    <None Include="img\m19y.jpg" />
     <Content Include="opencv_core220.dll">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
@@ -200,6 +234,12 @@
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
     <None Include="title.png" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\CA\ContourAnalysis.csproj">
+      <Project>{ddbd6cf7-2367-4084-a609-1b9c1885dc98}</Project>
+      <Name>ContourAnalysis</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
